memory leak $? fix
