+++ srcs/parsing/expander.c
@@ -3,49 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   expander.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: achabrer <achabrer@student.42porto.com>    +#+  +:+       +#+        */
+/*   By: jgomes-v <jgomes-v@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 17:49:43 by jgomes-v          #+#    #+#             */
-/*   Updated: 2024/01/30 11:11:38 by achabrer         ###   ########.fr       */
+/*   Updated: 2024/01/31 12:23:36 by jgomes-v         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-char	*get_key_expansion(char **temp)
-{
-	char	*start;
-	char	*key;
-
-	start = ++(*temp);
-	while (**temp && (ft_isalnum(**temp) || **temp == '_'))
-		(*temp)++;
-	key = ft_strdup(start);
-	key[*temp - start] = '\0';
-	(*temp)--;
-	return (key);
-}
-
-char	*append_value_to_content(char *new_content, char *value)
-{
-	size_t	old_len;
-
-	old_len = ft_strlen(new_content);
-	new_content = ft_realloc_str(new_content, old_len + ft_strlen(value) + 1);
-	ft_strlcat(new_content, value, old_len + ft_strlen(value) + 1);
-	return (new_content);
-}
-
-char	*append_char_to_content(char *new_content, char c)
-{
-	size_t	old_len;
-
-	old_len = ft_strlen(new_content);
-	new_content = ft_realloc_str(new_content, old_len + 2);
-	new_content[old_len] = c;
-	new_content[old_len + 1] = '\0';
-	return (new_content);
-}
 char	*expand_variable(char *cnt)
 {
 	char	*new;
@@ -88,7 +54,7 @@ char	*expand_variable(char *cnt)
 			cnt++;
 		else if (*cnt == '$' && *(cnt + 1) == '?' && !in_single_quotes)
 		{
-			new = append_value_to_content(new, ft_itoa(sh()->exit_status));
+			new = append_value_to_content_error(new, ft_itoa(sh()->exit_status));
 			cnt++;
 		}
 		else
